requirements.txt
Flask==1.1.2
flask_cors==3.0.10
flask_sqlalchemy==2.5.1
itemadapter==0.2.0
python_frontmatter==1.0.0
scrapy==2.5.0
